Added helpfull comments to explain classes and moved Color.java
